% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_partition.R
\name{append_partition}
\alias{append_partition}
\title{Function that given a partitioned dataset, a variable and a value partitions the dataset at specified cut point}
\usage{
append_partition(partition, variable, value)
}
\arguments{
\item{partition}{partitioned data}

\item{variable}{variable}

\item{value}{value}
}
\value{
list of partitions
}
\description{
Function that given a partitioned dataset, a variable and a value partitions the dataset at specified cut point
}
\examples{
partition(mtcars, 1, 20)

}
