% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_la_model}
\alias{build_la_model}
\title{Builds a look ahead model given a path obtained by the id_creator, trying the n-best split at any point and returns the loss of the corresponding model}
\usage{
build_la_model(
  path,
  model,
  max_tree_depth,
  min_leaf_size,
  measure,
  look_ahead,
  restr_la_var
)
}
\arguments{
\item{path}{path of splits to be exected}

\item{model}{main model}

\item{max_tree_depth}{maximal tree depth}

\item{min_leaf_size}{minimal leaf size}

\item{measure}{measure}

\item{look_ahead}{depth of looking ahead}

\item{restr_la_var}{whether or not the dataframe should be restricted to carry only one split per variable}
}
\value{
model loss
}
\description{
Builds a look ahead model given a path obtained by the id_creator, trying the n-best split at any point and returns the loss of the corresponding model
}
