% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tree.R
\name{build_tree}
\alias{build_tree}
\title{Function, that given data builds the next step}
\usage{
build_tree(data, target)
}
\arguments{
\item{data}{partitioned data}

\item{target}{target}
}
\value{
list of partitions
}
\description{
Function, that given data builds the next step
}
\examples{

}
