% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_model_loss}
\alias{calc_model_loss}
\title{Calvulates RMSE loss of model}
\usage{
calc_model_loss(model, measure)
}
\arguments{
\item{model}{model}

\item{measure}{measure}
}
\value{
loss
}
\description{
Calvulates RMSE loss of model
}
