% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_children}
\alias{find_children}
\title{Creates offspring and updates parent}
\usage{
find_children(node, split, measure)
}
\arguments{
\item{node}{node}

\item{split}{applied split}

\item{measure}{measure}
}
\value{
list with updated parent and both children
}
\description{
Creates offspring and updates parent
}
