% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_split.R
\name{find_split}
\alias{find_split}
\title{Function that given data and target variable returns frist optimal split point}
\usage{
find_split(data, target)
}
\arguments{
\item{data}{data}

\item{target}{target variable}
}
\value{
split point
}
\description{
Function that given data and target variable returns frist optimal split point
}
\examples{

}
