% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_la_split}
\alias{get_la_split}
\title{Gets the best split from all possible considered la-models}
\usage{
get_la_split(
  model,
  consider,
  look_ahead,
  max_tree_depth,
  min_leaf_size,
  measure,
  restr_la_var
)
}
\arguments{
\item{model}{input model}

\item{consider}{considered splits in each level}

\item{look_ahead}{depth of looking ahead}

\item{max_tree_depth}{maximal tree depth}

\item{min_leaf_size}{minimal leaf size}

\item{measure}{measure}

\item{restr_la_var}{whether or not the dataframe should be restricted to carry only one split per variable}
}
\value{
split number
}
\description{
Gets the best split from all possible considered la-models
}
