% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{id_creator}
\alias{id_creator}
\title{Translates decimal to ids of length look_ahead with range ! trough consider}
\usage{
id_creator(consider, look_ahead)
}
\arguments{
\item{consider}{considered amount of splits}

\item{look_ahead}{amount of future splits to be tried}
}
\value{
ids
}
\description{
Translates decimal to ids of length look_ahead with range ! trough consider
}
