% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{la_tree}
\alias{la_tree}
\title{Function that calculates the residual sum of squares}
\usage{
la_tree(
  data,
  target,
  task_type = "regr",
  look_ahead = 2,
  consider = 5,
  restr_la_var = TRUE,
  cap_la = 3,
  max_tree_depth = 6,
  min_leaf_size = 5,
  measure = "RMSE",
  prune_parameter = NULL
)
}
\arguments{
\item{data}{data}

\item{target}{target variable, eather numeric or as string}

\item{task_type}{task type}

\item{look_ahead}{hyperparameter how far to look into future splits}

\item{consider}{hyperparameter how many splits to consider in each future split}

\item{restr_la_var}{whether or not the dataframe should be restricted to carry only one split per variable}

\item{cap_la}{sets the maximum depth of the tree when look_ahead should still be applied}

\item{max_tree_depth}{hyperparameter to choose the maximal depth of the tree}

\item{min_leaf_size}{hyperparameter to choose the minimal size of a leaf allowed}

\item{measure}{measure}

\item{prune_parameter}{hyperparameter that is used for pruning the model}
}
\value{
model containing the used time, decission tree, preprocessing list and RMSE loss
}
\description{
Function that calculates the residual sum of squares
}
