% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data.R
\name{list_data}
\alias{list_data}
\title{Function that given a dataset transforms it into a list, unless it already is a list}
\usage{
list_data(data)
}
\arguments{
\item{data}{data}
}
\value{
data storred in a list
}
\description{
Function that given a dataset transforms it into a list, unless it already is a list
}
\examples{

}
