% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loss_nv}
\alias{loss_nv}
\title{Calculates loss of a split given name-value pair}
\usage{
loss_nv(measure, node, name, value)
}
\arguments{
\item{measure}{performance measure}

\item{node}{node}

\item{name}{variable name}
}
\value{
loss value
}
\description{
Calculates loss of a split given name-value pair
}
