% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loss_RSS}
\alias{loss_RSS}
\title{Calculates the residual sum of squares}
\usage{
loss_RSS(y)
}
\arguments{
\item{y}{dependent variable vector}
}
\value{
loss value
}
\description{
Calculates the residual sum of squares
}
