% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Function that given a dataset ant target variable creates a decission tree}
\usage{
main()
}
\value{
decission tree
}
\description{
Function that given a dataset ant target variable creates a decission tree
}
\examples{

}
