% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_X_numeric}
\alias{make_X_numeric}
\title{Makes all non numeric variables numeric for internal use}
\usage{
make_X_numeric(X, pp)
}
\arguments{
\item{X}{independent variables}

\item{pp}{preprocessing list that maps from factor to numeric values}
}
\value{
data frame containing only numeric variables
}
\description{
Makes all non numeric variables numeric for internal use
}
