% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Function that given a dataset, a variable and a value partitions the dataset at specified cut point}
\usage{
partition(data, split)
}
\arguments{
\item{data}{data}

\item{split}{list of variable and value}
}
\value{
list of partitions
}
\description{
Function that given a dataset, a variable and a value partitions the dataset at specified cut point
}
\examples{

}
