% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_la_tree}
\alias{predict_la_tree}
\title{Predicts the dependent variable based on model obtained from la_tree on new X data}
\usage{
predict_la_tree(newdata, model_list)
}
\arguments{
\item{newdata}{eather whole data or only independent variables}

\item{model_list}{model and preprocessing list from la_tree}
}
\value{
vector ob predictions
}
\description{
Predicts the dependent variable based on model obtained from la_tree on new X data
}
