% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict_obs}
\alias{predict_obs}
\title{Predicts individual observatons gets an id vector obtained from the node names in the model. This can be obtained by running (id_vec <- sapply(model, function(node) node$name)). This is needed to travel trough the tree}
\usage{
predict_obs(observation, model, id_vec)
}
\arguments{
\item{observation}{dataframe containing all X values of one observation}

\item{model}{model}

\item{id_vec}{vector of node id's}
}
\value{
estimate of dependen variable
}
\description{
Predicts individual observatons gets an id vector obtained from the node names in the model. This can be obtained by running (id_vec <- sapply(model, function(node) node$name)). This is needed to travel trough the tree
}
