% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sl_df}
\alias{sl_df}
\title{Creates a data frame storing name split and loss in a node}
\usage{
sl_df(node, min_leaf_size, measure)
}
\arguments{
\item{node}{node}

\item{min_leaf_size}{minimal allowed leaf size}

\item{measure}{measure}
}
\value{
df with name, value and loss
}
\description{
Creates a data frame storing name split and loss in a node
}
