% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sl_lists.R
\name{sl_lists}
\alias{sl_lists}
\title{Function that given a split_list returns the loss value to all corresponding splitting points}
\usage{
sl_lists(data, target)
}
\arguments{
\item{data}{data}

\item{target}{target variable}
}
\value{
two lists, one containing the splitting points, one containing the loss
}
\description{
Function that given a split_list returns the loss value to all corresponding splitting points
}
\examples{

}
