% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_list.R
\name{split_list}
\alias{split_list}
\title{Function that given a dataset and the target variable returns all possible split points}
\usage{
split_list(data, target)
}
\arguments{
\item{data}{data}

\item{target}{target variable}
}
\value{
matrix of possible spliting points
}
\description{
Function that given a dataset and the target variable returns all possible split points
}
\examples{

}
