% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_to_number.R
\name{target_to_number}
\alias{target_to_number}
\title{Function that given the target and data transforms the target to the column number in the data, when given as string}
\usage{
target_to_number(data, target)
}
\arguments{
\item{data}{data}

\item{target}{target}
}
\value{
target as numeric value
}
\description{
Function that given the target and data transforms the target to the column number in the data, when given as string
}
\examples{

}
