% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_type.R
\name{task_type}
\alias{task_type}
\title{Function that given a dataset and a target variable returns the task type}
\usage{
task_type(data, target)
}
\arguments{
\item{data}{dataset}

\item{target}{target variable}
}
\value{
task type
}
\description{
Function that given a dataset and a target variable returns the task type
}
\examples{

}
