% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{terminate_mtd}
\alias{terminate_mtd}
\title{Terminated a node, if maximal tree depth has been reached}
\usage{
terminate_mtd(node, max_tree_depth)
}
\arguments{
\item{node}{node}

\item{max_tree_depth}{maximal tree depth}
}
\value{
terminated node that is now a leaf
}
\description{
Terminated a node, if maximal tree depth has been reached
}
