% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{terminate_now}
\alias{terminate_now}
\title{Terminates a node and makes it a leaf without condition}
\usage{
terminate_now(node)
}
\arguments{
\item{node}{node}
}
\value{
terminated node
}
\description{
Terminates a node and makes it a leaf without condition
}
