% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{la_tree}
\alias{la_tree}
\title{Function that calculates the residual sum of squares}
\usage{
la_tree(
  data,
  target,
  task_type = "regr",
  look_ahead,
  consider = 5,
  max_tree_depth,
  min_leaf_size,
  measure = "RMSE",
  prune_parameter = NULL
)
}
\arguments{
\item{data}{data}

\item{target}{target variable, eather numeric or as string}

\item{task_type}{task type}

\item{look_ahead}{hyperparameter how far to look into future splits}

\item{consider}{hyperparameter how many splits to consider in each future split}

\item{max_tree_depth}{hyperparameter to choose the maximal depth of the tree}

\item{min_leaf_size}{hyperparameter to choose the minimal size of a leaf allowed}

\item{measure}{measure}

\item{prune_parameter}{hyperparameter that is used for pruning the model}
}
\value{
model containing the used time, decission tree, preprocessing list and RMSE loss
}
\description{
Function that calculates the residual sum of squares
}
