% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_function.R
\name{loss_function}
\alias{loss_function}
\title{Function that given a target variable, a partition and a task type returns a loss value}
\usage{
loss_function(partition, target, task_type = "regr")
}
\arguments{
\item{target}{target variable}

\item{task_type}{task type}

\item{partitionition}{partition}
}
\value{
list of partitionitions
}
\description{
Function that given a target variable, a partition and a task type returns a loss value
}
\examples{

}
